I would have thought that an Integer  Primary Key feild would already have the obvious automatic defautl value.

+++ FMTBackEnd/userNoUsername/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import AbstractUser
 # Create your models here.
 
 class UsernamelessUser(AbstractUser):
-    email = models.CharField(max_length = 255, unique = True)
-    url_exten = models.PositiveIntegerField(primary_key = True )
+    email = models.CharField( max_length = 255, unique = True)
+    url_exten = models.BigAutoField( primary_key = True )
     
     USERNAME_FIELD = "email"
     REQUIRED_FIELDS = []
